+++ tccelf.h
@@ -3,11 +3,11 @@
 /* put stab debug information */
 
 typedef struct {
-    unsigned long n_strx;         /* index into string table of name */
-    unsigned char n_type;         /* type of symbol */
-    unsigned char n_other;        /* misc info (usually empty) */
-    unsigned short n_desc;        /* description field */
-    unsigned long n_value;        /* value of symbol */
+    unsigned int n_strx;         /* index into string table of name */
+    unsigned char n_type;        /* type of symbol */
+    unsigned char n_other;       /* misc info (usually empty) */
+    unsigned short n_desc;       /* description field */
+    unsigned int n_value;        /* value of symbol */
 } Stab_Sym;
 
 int add_elf_sym(Section *s, unsigned long value, unsigned long size,
